// src/masters/inspection26/dto/create-inspection26.dto.ts
export class CreateInspection26Dto {
  address?: string;
  city?: string;
  stzip?: string;
  date?: string;
  fileno?: string;
  appraiser?: string;
  borrower?: string;
  timein?: string;
  timeout?: string;
  reporttype?: string;
  occupant?: string;
  dwelling_style?: string;
  structure?: string;
  units?: string;
  stories?: string;
  yearbuilt?: string;
  age?: string;
  effage?: string;
  rel?: string;
  pud?: string;
  hoa?: string;
  hoafreq?: string;
  condoproj?: string;
  status?: string;

  site?: any;
  exterior?: any;
  interior?: any;
  basement?: any;
  measurements?: any;
}